move KSPGradleLogger and KSPLoader into aa module
